CLN: create core/sorting.py
just a small reorg to put sorting / grouping utilities into a separate
area

Author: Jeff Reback <jeff@reback.net>

Closes #15402 from jreback/sorting and squashes the following commits:

fdcf9a1 [Jeff Reback] change a couple of sorting.py functions to be non-private (public to pandas internals)
90ff22d [Jeff Reback] split up some value_counts groupby tests a bit
18ea902 [Jeff Reback] CLN: create core/sorting.py
92dcb07 [Jeff Reback] CLN: remove numpy_groupby as not used
